Simplified by using a table view controller and reloading on app resume.
By using a table view controller the keyboard is handled automatically
which significantly simplifies the class. The settings table was
removed and all keyboard notifications removed. Instead it uses the
built-in table.

Now it handles the case of leaving the app and changing a setting in
preferences, when resuming the app the settings will reload to show the
change.